+++ packages/ui-components/web/src/index.ts
@@ -6,6 +6,7 @@ export {
   DesktopFlyOutModalStyles,
   MOBILE_FULL_SCREEN_CUSTOM_STYLE,
   DesktopOverlayModalStyles,
+  ACTION_COLORS_CSS_VAR,
 } from './consts';
 
 // Modals