+++ packages/plugin-image/web/src/toolbar/image-settings.js
@@ -12,6 +12,7 @@ import {
   Loader,
   SettingsMobileHeader,
   SettingsSeparator,
+  SettingsPanelHeader,
 } from 'wix-rich-content-ui-components';
 import styles from '../../statics/styles/image-settings.scss';
 import { DIVIDER } from '../consts';
@@ -24,7 +25,6 @@ class ImageSettings extends Component {
     const { t, theme } = props;
     this.styles = mergeStyles({ styles, theme });
     this.updateLabel = t('ImageSettings_Update');
-    this.headerText = t('ImageSettings_Header');
     this.captionLabel = t('ImageSettings_Caption_Label');
     this.captionInputPlaceholder = t('ImageSettings_Caption_Input_Placeholder');
     this.altLabel = t('ImageSettings_Alt_Label');
@@ -185,7 +185,10 @@ class ImageSettings extends Component {
             t={t}
           />
         ) : (
-          <h3 className={this.styles.imageSettingsTitle}>{this.headerText}</h3>
+          <SettingsPanelHeader
+            title={t('ImageSettings_Header')}
+            onClose={this.revertComponentData}
+          />
         )}
         <div
           className={classNames(styles.imageSettings_scrollContainer, {