+++ packages/tiptap-extensions/web/src/tiptap-extensions.ts
@@ -13,10 +13,6 @@ import Heading from '@tiptap/extension-heading';
 import Blockquote from '@tiptap/extension-blockquote';
 import OrderedList from '@tiptap/extension-ordered-list';
 import ListItem from '@tiptap/extension-list-item';
-import { Paragraph } from './paragraph';
-import { createBold } from './bold';
-import { createBulletedList } from './bulleted-list';
-import { createSpoiler } from './spoiler';
 import headingDataDefaults from 'ricos-schema/dist/statics/heading.defaults.json';
 import { Attributes } from '@tiptap/react';
 import styles from './statics/styles.scss';
@@ -46,14 +42,10 @@ export const tiptapExtensions: AnyExtension[] = [
   Italic,
   ListItem,
   OrderedList,
-  Paragraph,
   Text,
-  createBulletedList(),
-  createBold(),
   CodeBlock.configure({
     HTMLAttributes: {
       class: styles.code,
     },
   }),
-  createSpoiler(),
 ];